--- src/main/java/com/google/api/generator/gapic/composer/samplecode/ServiceClientSampleUtil.java
@@ -1,117 +1,0 @@
-// Copyright 2022 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package com.google.api.generator.gapic.composer.samplecode;
-
-import com.google.api.generator.engine.ast.AssignmentExpr;
-import com.google.api.generator.engine.ast.Expr;
-import com.google.api.generator.engine.ast.MethodInvocationExpr;
-import com.google.api.generator.engine.ast.TypeNode;
-import com.google.api.generator.engine.ast.Variable;
-import com.google.api.generator.engine.ast.VariableExpr;
-import com.google.api.generator.gapic.composer.defaultvalue.DefaultValueComposer;
-import com.google.api.generator.gapic.model.MethodArgument;
-import com.google.api.generator.gapic.model.ResourceName;
-import com.google.api.generator.gapic.utils.JavaStyle;
-import com.google.common.base.Preconditions;
-import java.util.Collections;
-import java.util.List;
-import java.util.Map;
-import java.util.function.Function;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
-
-public class ServiceClientSampleUtil {
-  // Create a list of RPC method arguments' variable expressions.
-  static List<VariableExpr> createRpcMethodArgumentVariableExprs(List<MethodArgument> arguments) {
-    return arguments.stream()
-        .map(
-            arg ->
-                VariableExpr.withVariable(
-                    Variable.builder()
-                        .setName(JavaStyle.toLowerCamelCase(arg.name()))
-                        .setType(arg.type())
-                        .build()))
-        .collect(Collectors.toList());
-  }
-
-  // Create a list of RPC method arguments' default value expression.
-  static List<Expr> createRpcMethodArgumentDefaultValueExprs(
-      List<MethodArgument> arguments, Map<String, ResourceName> resourceNames) {
-    List<ResourceName> resourceNameList =
-        resourceNames.values().stream().collect(Collectors.toList());
-    Function<MethodArgument, MethodInvocationExpr> stringResourceNameDefaultValueExpr =
-        arg ->
-            MethodInvocationExpr.builder()
-                .setExprReferenceExpr(
-                    DefaultValueComposer.createResourceHelperValue(
-                        resourceNames.get(arg.field().resourceReference().resourceTypeString()),
-                        arg.field().resourceReference().isChildType(),
-                        resourceNameList,
-                        arg.field().name()))
-                .setMethodName("toString")
-                .setReturnType(TypeNode.STRING)
-                .build();
-    return arguments.stream()
-        .map(
-            arg ->
-                !isStringTypedResourceName(arg, resourceNames)
-                    ? DefaultValueComposer.createMethodArgValue(
-                        arg, resourceNames, Collections.emptyMap(), Collections.emptyMap())
-                    : stringResourceNameDefaultValueExpr.apply(arg))
-        .collect(Collectors.toList());
-  }
-
-  // Create a list of assignment expressions for variable expr with value expr.
-  static List<Expr> createAssignmentsForVarExprsWithValueExprs(
-      List<VariableExpr> variableExprs, List<Expr> valueExprs) {
-    Preconditions.checkState(
-        variableExprs.size() == valueExprs.size(),
-        "Expected the number of method arguments to match the number of default values.");
-    return IntStream.range(0, variableExprs.size())
-        .mapToObj(
-            i ->
-                AssignmentExpr.builder()
-                    .setVariableExpr(variableExprs.get(i).toBuilder().setIsDecl(true).build())
-                    .setValueExpr(valueExprs.get(i))
-                    .build())
-        .collect(Collectors.toList());
-  }
-
-  // Assign client variable expr with create client.
-  // e.g EchoClient echoClient = EchoClient.create()
-  static AssignmentExpr assignClientVariableWithCreateMethodExpr(VariableExpr clientVarExpr) {
-    return AssignmentExpr.builder()
-        .setVariableExpr(clientVarExpr.toBuilder().setIsDecl(true).build())
-        .setValueExpr(
-            MethodInvocationExpr.builder()
-                .setStaticReferenceType(clientVarExpr.variable().type())
-                .setReturnType(clientVarExpr.variable().type())
-                .setMethodName("create")
-                .build())
-        .build();
-  }
-
-  private static boolean isStringTypedResourceName(
-      MethodArgument arg, Map<String, ResourceName> resourceNames) {
-    return arg.type().equals(TypeNode.STRING)
-        && arg.field().hasResourceReference()
-        && resourceNames.containsKey(arg.field().resourceReference().resourceTypeString());
-  }
-
-  static boolean isProtoEmptyType(TypeNode type) {
-    return type.reference().pakkage().equals("com.google.protobuf")
-        && type.reference().name().equals("Empty");
-  }
-}
